Don't attempt to call MapUpdateSnapshot in recovery.
CreateUpdateSnapshots() is expected to fail in recovery most of the
time when updates are too large and we need space from /data. When it does we are supposed to fall back to an in-place update.
However because target_supports_snapshot_ is still true, we accidentally
call MapUpdateSnapshot(). This happens to fail to create dm-snapshot
targets and winds up with only a dm-linear target, so the in-place
update proceeds normally.

When disabling dm-snapshot, MapUpdateSnapshot explicitly fails much
earlier, causing sideload to fail.

Fix this by setting a new prop which forces update_engine into using CreateLogicalPartitions instead(direct writes to the logical partition rather than a COW device).

Bug: 416748567
Test: adb sideload ota.zip
Change-Id: I5acd52c98655589f49e871fb26204a0517eb36fc