Do not send kUserCancelled status if action is cancelled internally
Sometimes, update-engine cancels PostinstallAction internally to service
other requests. This is only done when the PostinstallAction is running
optional postinstall scripts, and ensures that critical API calls such
as slot switching can be serviced. In addition, systemserver might
trigger postinstall script in the background, which is something
OTA clients are not aware of(nor they should). Hence we do not send
kUserCancelled status in this case, to keep OTA client unaware of
async background postinstall scripts.

Since the cancellation does not come from
caller, do not send kUserCancelled status.

Test: install update, trigger async dexopt, do slot switch
Bug: 438709490
Change-Id: Iaeae414853f1bee3ecf0678a4b2baf18738ca6b3