update_engine_unittests: use ASSERT_* in xz_extent_writer_unittest
Use ASSERT_* macros instead of EXPECT_* macros in the
xz_extent_writer_unittest.cc file. This ensures that the test will stop
execution immediately if an assertion fails, which is the desired
behavior for these tests. This change also makes the test consistent with
the zstd_extent_writer_unittest.cc test.

Bug: 436177354
Test: th, update_engine_unittest, update_engine_host_unittests
Change-Id: I5e0a6a5439b7ebff5d605786c4d2520fd12fe342
Signed-off-by: Sandeep Dhavale <dhavale@google.com>